Updated README and renamed calib.py to g27calib.py. Created spec file for pyinstaller and built exe.
